Uploaded dataset and midway through plotting of data
Dataset.xlsx contains the dataset we're working with, it requires an import on every re-opening of Matlab.

4dPlot.asv is the file used to plot the dataset. Currently the objective is to plot the four parameters.

sepal length, sepal width and petal length is to be respresented with world cordernates. (X,Y,Z) petalWidth with colour (black for 0, white for 3.5) and for the flower type the shape of the plotted point will differ for each flower respectively.